+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: misi-moh <misi-moh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/15 20:29:46 by misi-moh          #+#    #+#             */
-/*   Updated: 2022/12/15 20:35:32 by misi-moh         ###   ########.fr       */
+/*   Updated: 2022/12/31 13:00:31 by misi-moh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t n)
 {
 	char	*a;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	a = s;
